LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/uncontrolled-page.https-expected.txt


FAIL Fetch events should not go through uncontrolled page. assert_unreached: assert_equals: expected "a simple text file\n" but got "ERROR" Reached unreachable code
